Let a committed trader cancel, and keep action prices on the tick grid
Two defects in the simulator's action path, both at edges the suite did
not visit.

Trader._order_approved ran the same cash check for every order type, so a
cancel was checked against its own irrelevant size and price. A trader
whose cash was fully escrowed in resting orders was refused the cancel
that would have freed it, silently, with num_rejected_step incremented; a
modify that shrank or re-priced an order was refused the same way because
the check compared against cash before cancel_cash_transfer had returned
the old order's escrow. A cancel is now approved once the NAV gate passes,
and a modify or same-price limit upsert may spend the escrow the replaced
order releases. Seven tests in test_cash_check.py.

_set_price read resting prices out of a float32 snapshot and added its
offset in float, and the book keys its price map on Decimal(str(price))
with no rounding, so on any non-integer tick every re-quote at an occupied
level opened a new price level one ulp away and Trader._get_order_ID,
comparing Decimal against float, never found the trader's own order.
agg_LOB_raw is float64, _set_price snaps to the tick grid in Decimal, and
_get_order_ID compares as Decimal(str(price)). test_tick_grid.py, 14 tests.

Also: unused locals and one f-string flagged by pyflakes, a stale
observation_layout note in tunable_constants.json, and the documentation
set brought to the tree: doc/15 S2-13 and S3-4, doc/16 16.17, doc/17 44,
doc/10 counts and gaps, doc/04, 06, 13, 18, and a new operator runbook at
doc/26 linked from the README. Suite: 935 unit + 153 integration, green.

+++ gym_continuousDoubleAuction/envs/exchg/state_helper.py
@@ -529,8 +529,12 @@ class State_Helper(object):
                     ask_size_list[k] = -set[1].volume
                 else:
                     break
-        # Raw unnormalized snapshot
-        flattened_raw = np.concatenate([bid_price_list, bid_size_list, ask_price_list, ask_size_list]).astype(np.float32)
+        # Raw unnormalized snapshot. float64, not float32: `_set_price`
+        # reads resting prices out of this array to quote at a book level, and
+        # float32 cannot hold 100.1 (it reads back 100.0999984741211). The
+        # emitted observation is cast to float32 at emission, in `_stack`; the
+        # raw array is the one consumer that needs the price exact.
+        flattened_raw = np.concatenate([bid_price_list, bid_size_list, ask_price_list, ask_size_list]).astype(np.float64)
         self.agg_LOB_raw = flattened_raw
 
         # Calculate Level 1 midpoint price M. Both this and `mid_price` read
